requirements.txt
# Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Image Processing
Pillow==10.0.0

# 3D Processing
torch==2.1.0

# Development & Debugging
pytest==7.4.4
python-dotenv==1.0.0

# Optional but recommended for better performance
numba==0.58.1
llvmlite==0.41.1

# 3D Processing
torch==2.1.0
trimesh==4.0.5
pyrender==0.1.45
numpy==1.24.3

# 3D Plotting
matplotlib==3.7.1